requirements.txt
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
bcrypt==3.1.3
boto==2.47.0
boto3==1.4.4
botocore==1.5.75
certifi==2017.4.17
cffi==1.10.0
characteristic==14.3.0
chardet==3.0.4
constantly==15.1.0
cryptography==1.9
cssselect==1.0.1
DateTime==4.2
docutils==0.13.1
Fabric==1.13.2
Fabric3==1.13.1.post1
hyperlink==17.2.1
idna==2.5
incremental==17.5.0
jmespath==0.9.3
lxml==3.8.0
numpy==1.13.1
paramiko==2.2.1
parsel==1.2.0
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycparser==2.17
PyDispatcher==2.0.5
pymongo==3.4.0
PyNaCl==1.1.2
pyOpenSSL==17.0.0
python-dateutil==2.6.0
python-scrapyd-api==2.0.1
pytz==2017.2
queuelib==1.4.2
requests==2.18.1
s3transfer==0.1.10
Scrapy==1.4.0
scrapyd==1.2.0
scrapyd-client==1.1.0
service-identity==17.0.0
six==1.10.0
Twisted==17.5.0
urllib3==1.21.1
w3lib==1.17.0
zope.interface==4.4.2
